fix: unmarshaling blob structs.
Different docker registries might return different media types and we should be able to deal with them.
